Implement RTT calculation for candidate pair stats
Implement CurrentRoundTripTime, TotalRoundTripTime, RequestsSent, and
ResponsesReceived for candidate pair statistics.

This change on its own is not quite enough to get stats working end to
end. Notably, `(*Agent).GetSelectedCandidatePair()` copies both
candidates, which due to internally marshaling and unmarshaling the
candidate to implement the copy, ends up generating new candidate ids.

Because pion/webrtc stores stats in a map based on candidate ids, this
means that it is currently impractical to actually access these stats.